separate XMLBuilder logic as newfile
